backend/requirements.txt
fastapi
uvicorn[standard]
groq
requests
python-dotenv
pydantic
pandas
openpyxl